pyproject.toml
[project]
name = "ollama-deep-researcher"
version = "0.0.1"
description = "Fully local web research and summarization assistant with Ollama and LangGraph."
authors = [
    { name = "Lance Martin" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.55",
    "langchain-community>=0.3.9",
    "tavily-python>=0.5.0",
    "langchain-ollama>=0.3.6",
    "duckduckgo-search>=7.3.0",
    "httpx>=0.28.1",
    "markdownify>=0.11.0",
    "python-dotenv==1.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-json-logger>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.6.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ollama_deep_researcher"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
addopts = "-v -s --tb=short"

[tool.black]
target-version = ['py312']
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.uv-cache
    | \.venv
    | venv
    | env
    | build
    | dist
    | .*/(bin|lib|share|etc)
    | .*/pyvenv\.cfg
)/
'''

[tool.ruff]
target-version = "py312"
exclude = [
    "__pycache__",
    "migrations",
    "db.sqlite3",
    "*.pyc",
    ".uv-cache",
    ".venv",
    "venv",
    "env",
    "*/bin/**",
    "*/lib/**",
    "*/share/**",
    "*/etc/**",
    "*/pyvenv.cfg"
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] # E501 (line too long) is ignored